refactor: reorganize and consolidate Atlas cohort definition types and utilities
