+++ ui/apps/vue-mri-ui-lib/src/query-filter/components/QueryFilterModern.vue
@@ -27,12 +27,8 @@ import type {
   ConceptSetDetails,
   CreateConceptSetRequest,
 } from '../types/ConceptSetTypes'
-import type {
-  AtlasCohortDefinition,
-  ConceptSet,
-  CriteriaGroup,
-  CriteriaListItem,
-} from '../models/AtlasCohortDefinition'
+import type { AtlasCohortDefinition, ConceptSet, CriteriaGroup, CriteriaListItem } from '../types/AtlasTypes'
+
 import {
   loadConceptSets as apiLoadConceptSets,
   loadConceptSetDetails as apiLoadConceptSetDetails,
